fix: validate vacancy uuid explicitly and widen job direction id to int64
- return domain.ErrInvalidParam from UpdateVacancy/DeleteVacancy when the
  passed id fails uuid parsing, instead of silently falling through to the
  nullable-filter SQL guard
- rename paramInt32 -> paramInt64 and switch job direction ids to int64
  throughout repo/service/transport layers
- add empty-id checks and clearer 400 error messages on vacancy and job
  direction handlers

+++ internal/service/services.go
@@ -33,11 +33,11 @@ type Vacancies interface {
 	Delete(ctx context.Context, vacancyId string) error
 
 	ListJd(ctx context.Context, limit, offset int32) (*domain.JobDirections, error)
-	GetJd(ctx context.Context, jdId int32) (*domain.JobDirection, error)
+	GetJd(ctx context.Context, jdId int64) (*domain.JobDirection, error)
 
 	CreateJd(ctx context.Context, params *domain.CreateJobDirectionInput) (*domain.JobDirection, error)
-	UpdateJd(ctx context.Context, jdId int32, updJd *domain.UpdateJobDirectionInput) (*domain.JobDirection, error)
-	DeleteJd(ctx context.Context, jdId int32) error
+	UpdateJd(ctx context.Context, jdId int64, updJd *domain.UpdateJobDirectionInput) (*domain.JobDirection, error)
+	DeleteJd(ctx context.Context, jdId int64) error
 
 	Respond(ctx context.Context, vacancyId string, applicantsForm *domain.ApplicantsFormInput, fileInput *domain.FileUploadInput) error
 	GetRespondVacancies(ctx context.Context) (*domain.RespondVacancies, error)
